Revert "move api keys table to its own component"
This reverts commit 6d8fefaa786e958fec54fe7931a905e8c001b844.
