--- app/src/pages/settings/APIKeysTable.tsx
@@ -1,142 +1,0 @@
-import React, { useMemo } from "react";
-import {
-  ColumnDef,
-  flexRender,
-  getCoreRowModel,
-  useReactTable,
-} from "@tanstack/react-table";
-
-import { Flex, Icon, Icons } from "@arizeai/components";
-
-import { TextCell } from "@phoenix/components/table";
-import { tableCSS } from "@phoenix/components/table/styles";
-import { TableEmpty } from "@phoenix/components/table/TableEmpty";
-import { TimestampCell } from "@phoenix/components/table/TimestampCell";
-
-import { DeleteAPIKeyButton } from "./DeleteAPIKeyButton";
-import { APIKey } from "./types";
-export function APIKeysTable({
-  data,
-  handleDelete,
-}: {
-  data: APIKey[];
-  handleDelete: (id: string) => void;
-}) {
-  const columns = useMemo(() => {
-    const cols: ColumnDef<APIKey>[] = [
-      {
-        header: "Name",
-        accessorKey: "name",
-      },
-      {
-        header: "Description",
-        accessorKey: "description",
-        cell: TextCell,
-      },
-      {
-        header: "Created At",
-        accessorKey: "createdAt",
-        cell: TimestampCell,
-      },
-      {
-        header: "Expires At",
-        accessorKey: "expiresAt",
-        cell: TimestampCell,
-      },
-      {
-        header: "",
-        accessorKey: "id",
-        size: 10,
-        cell: ({ row }) => {
-          return (
-            <Flex direction="row" justifyContent="end" width="100%">
-              <DeleteAPIKeyButton
-                handleDelete={() => {
-                  handleDelete(row.original.id);
-                }}
-              />
-            </Flex>
-          );
-        },
-        meta: {
-          textAlign: "right",
-        },
-      },
-    ];
-    return cols;
-  }, [handleDelete]);
-  const table = useReactTable<APIKey>({
-    columns,
-    data,
-    getCoreRowModel: getCoreRowModel(),
-  });
-  const rows = table.getRowModel().rows;
-  const isEmpty = table.getRowModel().rows.length === 0;
-  return (
-    <table css={tableCSS}>
-      <thead>
-        {table.getHeaderGroups().map((headerGroup) => (
-          <tr key={headerGroup.id}>
-            {headerGroup.headers.map((header) => (
-              <th colSpan={header.colSpan} key={header.id}>
-                {header.isPlaceholder ? null : (
-                  <div
-                    {...{
-                      className: header.column.getCanSort()
-                        ? "cursor-pointer"
-                        : "",
-                      onClick: header.column.getToggleSortingHandler(),
-                      style: {
-                        left: header.getStart(),
-                        width: header.getSize(),
-                      },
-                    }}
-                  >
-                    {flexRender(
-                      header.column.columnDef.header,
-                      header.getContext()
-                    )}
-                    {header.column.getIsSorted() ? (
-                      <Icon
-                        className="sort-icon"
-                        svg={
-                          header.column.getIsSorted() === "asc" ? (
-                            <Icons.ArrowUpFilled />
-                          ) : (
-                            <Icons.ArrowDownFilled />
-                          )
-                        }
-                      />
-                    ) : null}
-                  </div>
-                )}
-              </th>
-            ))}
-          </tr>
-        ))}
-      </thead>
-      {isEmpty ? (
-        <TableEmpty message="No Keys" />
-      ) : (
-        <tbody>
-          {rows.map((row) => {
-            return (
-              <tr key={row.id}>
-                {row.getVisibleCells().map((cell) => {
-                  return (
-                    <td key={cell.id}>
-                      {flexRender(
-                        cell.column.columnDef.cell,
-                        cell.getContext()
-                      )}
-                    </td>
-                  );
-                })}
-              </tr>
-            );
-          })}
-        </tbody>
-      )}
-    </table>
-  );
-}
