--- app/src/pages/settings/types.tsx
@@ -1,7 +1,0 @@
-export interface APIKey {
-  readonly id: string;
-  readonly name: string;
-  readonly description?: string | null;
-  readonly createdAt: string;
-  readonly expiresAt?: string | null;
-}
